Remove codec copy
When runtime.exec() runs for the commad, the terminal is supposed to tell
whether the operation succeeds or not. This program uses the Process#waitFor()
method to check this, but the convertion with codec copy doesn't respond its
failure even though it indeed fails. This problem cannot be solved for now.
The codec copy has to be removed.